refactor: split errors into separate files, delete unused ones
