+++ src/credential/presentation/errors.ts
@@ -1,0 +1,49 @@
+import { IoWalletError, serializeAttrs } from "../../utils/errors";
+
+/**
+ * An error subclass thrown when auth request decode fail
+ *
+ */
+export class AuthRequestDecodeError extends IoWalletError {
+  static get code(): "ERR_IO_WALLET_AUTHENTICATION_REQUEST_DECODE_FAILED" {
+    return "ERR_IO_WALLET_AUTHENTICATION_REQUEST_DECODE_FAILED";
+  }
+
+  code = "ERR_IO_WALLET_AUTHENTICATION_REQUEST_DECODE_FAILED";
+
+  /** The Claim for which the validation failed. */
+  claim: string;
+
+  /** Reason code for the validation failure. */
+  reason: string;
+
+  constructor(
+    message: string,
+    claim: string = "unspecified",
+    reason: string = "unspecified"
+  ) {
+    super(serializeAttrs({ message, claim, reason }));
+    this.claim = claim;
+    this.reason = reason;
+  }
+}
+
+/**
+ * When selecting a public key from an entity configuration, and no one meets the requirements for the scenario
+ *
+ */
+export class NoSuitableKeysFoundInEntityConfiguration extends Error {
+  static get code(): "ERR_NO_SUITABLE_KEYS_NOT_FOUND" {
+    return "ERR_NO_SUITABLE_KEYS_NOT_FOUND";
+  }
+
+  code = "ERR_NO_SUITABLE_KEYS_NOT_FOUND";
+
+  /**
+   * @param scenario describe the scenario in which the error arise
+   */
+  constructor(scenario: string) {
+    const message = `Entity configuration do not provide any suitable keys (${scenario}).`;
+    super(message);
+  }
+}
